Make mybatis database connection configurable, move bedload sediment calculation to its own class and call from autoproc.  add tests.
