+++ src/main/resources/mybatis/mybatis.conf.xml
@@ -20,12 +20,21 @@
 		<typeAlias alias="TimeSeriesRecord" type="gov.usgs.wma.gcmrc.model.TimeSeriesRecord"/>
 	</typeAliases>
 	
-	<environments default="development">
-		<environment id="development">
+	<environments default="oracle">
+		<environment id="postgres">
 			<transactionManager type="JDBC"/>
 			<dataSource type="POOLED">
 				<property name="driver" value="org.postgresql.Driver"/>
 				<property name="url" value="jdbc:postgresql://${gdaws.dbHost}:${gdaws.dbPort}/${gdaws.dbName}?currentSchema=${gdaws.schemaName}"/>
+				<property name="username" value="${gdaws.dbUser}"/>
+				<property name="password" value="${gdaws.dbPwd}"/>
+			</dataSource>
+		</environment>
+		<environment id="oracle">
+			<transactionManager type="JDBC"/>
+			<dataSource type="POOLED">
+				<property name="driver" value="oracle.jdbc.OracleDriver"/>
+				<property name="url" value="jdbc:oracle:thin:@${gdaws.dbHost}:${gdaws.dbPort}:${gdaws.dbName}"/>
 				<property name="username" value="${gdaws.dbUser}"/>
 				<property name="password" value="${gdaws.dbPwd}"/>
 			</dataSource>
